+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+# v2.10.37 (Thu May 19 2022)
+
+#### 🐛 Bug Fix
+
+- fix(sidebar): fix route for navigation [#60](https://github.com/AppQuality/unguess-design-system/pull/60) ([@marcbon](https://github.com/marcbon))
+
+#### Authors: 1
+
+- Marco Bonomo ([@marcbon](https://github.com/marcbon))
+
+---
+
+# v2.10.36 (Thu May 19 2022)
+
+#### 🐛 Bug Fix
+
+- Cup 721 service cards + info cards [#58](https://github.com/AppQuality/unguess-design-system/pull/58) ([@marcbon](https://github.com/marcbon))
+
+#### Authors: 1
+
+- Marco Bonomo ([@marcbon](https://github.com/marcbon))
+
+---
+
+# v2.10.35 (Wed May 18 2022)
+
+#### 🐛 Bug Fix
+
+- feat(sidebar): add templates menu item [#59](https://github.com/AppQuality/unguess-design-system/pull/59) ([@cannarocks](https://github.com/cannarocks))
+
+#### Authors: 1
+
+- Luca Cannarozzo ([@cannarocks](https://github.com/cannarocks))
+
+---
+
 # v2.10.34 (Wed May 18 2022)
 
 #### 🐛 Bug Fix